// apps/frontend-ng/src/app/core/api/generated/models/user.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface User {
  '_count'?: {
'devices'?: number;
'alerts'?: number;
};
  createdAt?: string;
  customerId?: string | null;
  email?: string;
  id?: string;
  role?: 'USER' | 'ADMIN' | 'SUPERADMIN' | 'READONLY';
  username?: string;
}